bdpt fixes; exr and png exports; exr imports for comparison stats; blend mode and normal flipping on thin opaque surfaces

+++ PathTracing/Metal/Utility.metal
@@ -16,13 +16,13 @@ using namespace raytracing;
 void cameraRayPDF(constant Uniforms& uniforms, float3 w, thread float& positionPDF, thread float& directionPDF) {
     positionPDF = 1.0f;
     float cosCamera = dot(w, uniforms.camera.forward);
-    float A = 4.0f * (uniforms.width / uniforms.height) * pow(tan(M_PI_F * CAMERA_FOV_ANGLE * 0.5f / 180.0f), 2.0f);
+    float A = 4.0f * (static_cast<float>(uniforms.width) / static_cast<float>(uniforms.height)) * pow(tan(M_PI_F * CAMERA_FOV_ANGLE * 0.5f / 180.0f), 2.0f);
     directionPDF = 1.0f / (A * pow(cosCamera, 3.0f));
 }
 
 float3 cameraWe(constant Uniforms& uniforms, float3 position) {
     float3 w = normalize(position - uniforms.camera.position);
-    float A = 4.0f * (uniforms.width / uniforms.height) * pow(tan(M_PI_F * CAMERA_FOV_ANGLE * 0.5f / 180.0f), 2.0f);
+    float A = 4.0f * (static_cast<float>(uniforms.width) / static_cast<float>(uniforms.height)) * pow(tan(M_PI_F * CAMERA_FOV_ANGLE * 0.5f / 180.0f), 2.0f);
     return 1.0f / (A * pow(dot(uniforms.camera.forward, w), 4.0f));
 }
